Notes from Meeting on 11/20/19
See notes from Meeting on 11/20/19:
(1) Be consistent with TRUE and T
(2) no need to do dtm or lda JSON functions in tester
(3) be sure to call with every parameter in tester
(4) use validate after json function to show it returns valid json
(5) for plots return path of saved pdf; update wordcloud functions in api to do this; but remove worldcloud functions with JSON in tester.
(6) remeber best.k will take a while to run.